pyproject.toml
[project]
name = "unbound-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "RameshArvind", email = "ramesharvind1994@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "daytona>=0.115.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
unbound-llm = "unbound_llm:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
